Having an abstract test class did not work as expected :(
